// types.ts
export enum TabType {
  BASIC = 'basic',
  CROP = 'crop',
  TEXT = 'text',
  BEAUTY = 'beauty',
  BODY = 'body',
  FILTER = 'filter',
  EFFECT = 'effect',
  AI_PRO = 'ai_pro'
}